requirements.txt
python-telegram-bot==21.3
httpx==0.27.0
Pillow==10.3.0
python-dotenv==1.0.1